feat(FallbackExceptionHandler): extract handleAny to FallbackExceptionHandler
FallbackExceptionHandler has lowest precedence and won't catch spring built-in exceptions

Closes #23
